Fixed error caused by not filling whole cms tree
Caused by calling filteredChildren instead of childNodes()
the tree was filtered while building the sitetree.
If the sitetree does not contain for example a page the
user cannot access, there ist no route to this page.
The result than is an HttpNotFoundException and there is
no control to route the user to a login page or something
comparable.
So even if the navigation of a site is filtered by
filteredChildren the sitetree has to be filled completely