Finalized actions framework
Actions are now REST-like devided in items and collections. So you
assign actions to an existing item or a collection.
Then on your "editor-page" to asc in your search results for item
actions and globally around your search-result for collection actions.
Also added onClick property for actions and setted enabled default true.