Huge rewite of routing stack
From now on Cmsable will rewrite the path of the original request to an
internal laravel route.
This means, you have to register every controller as you would normally
do in routes.php and then map your pageType to one of this URLs. The
only thing cmsable does is then:
url > pageType > replace url with pageType->targetPath

Also the whole register-process was completely rewritten because now
cmsable acts as a http-middleware to laravel (in l4 with stacked
implementation)