Added support for parameters in breadcrumbs
Luckily a very simple method solved a lot of breadcrumb headache for me.
Now it is possible to store any parameters with the current page in
session. So if you are for example on a search page and then on a detail
page, you will get exactly your search back if you click the parent
breadcrumb yay :-)