Implemented PageType routing
PageType routing is the opposite routing of cmsable standard routing. If
no page with the path was found in db, a second routing method will
fire.
You can pass an array of Routenames to your pagetype. If cmsable finds
exactly one page which pagetype has the current routename in its
routenames, it will call your controller creator and provide the config
for it. This allows to have configurable subpages within one page. So if
you need for example a password-email and a password-reset after
emailing page you only need one cms page.