Fixed many erros caused by request replacing
The middleware now injects the CmsRequest, the Cms\Application can
create a CmsRequest on demand if none were created during routing. But
if it has to create a CmsRequest it is not published by event, because
this is only a local request.
Also updated the facades and laravel app when request is replaced