Allow manually passing the resource name of validators
This commit allows to really get a validator via
validator($resourceName). The old method didnt really allow to manually
pass the resource name. Other methods get called and asked the
distributor for an auto detection. This made it impossible to use
ResourceValidator objects on console, where no request is present.
The same has to be done with other resource dependent objects...